SDK map update (full refactor)
