+++ app/src/main/java/pl/llp/aircasting/screens/new_session/choose_location/ChooseLocationViewMvcImpl.kt
@@ -6,8 +6,8 @@ import android.view.ViewGroup
 import android.widget.Button
 import androidx.fragment.app.FragmentManager
 import com.google.android.gms.common.api.Status
-import com.google.android.libraries.maps.*
-import com.google.android.libraries.maps.model.LatLng
+import com.google.android.gms.maps.*
+import com.google.android.gms.maps.model.LatLng
 import com.google.android.libraries.places.api.model.Place
 import com.google.android.libraries.places.widget.AutocompleteSupportFragment
 import com.google.android.libraries.places.widget.listener.PlaceSelectionListener
@@ -94,13 +94,6 @@ class ChooseLocationViewMvcImpl : BaseObservableViewMvc<ChooseLocationViewMvc.Li
         }
     }
 
-    override fun onMapReady(googleMap: GoogleMap?) {
-        googleMap ?: return
-        mMap = googleMap
-        setZoomPreferences()
-        resetMapToDefaults()
-    }
-
     override fun onDestroy() {
         mMap = null
         mMapFragment?.onDestroy()
@@ -146,4 +139,11 @@ class ChooseLocationViewMvcImpl : BaseObservableViewMvc<ChooseLocationViewMvc.Li
 
         return mapOptions
     }
+
+    override fun onMapReady(googleMap: GoogleMap) {
+        mMap = googleMap
+        setZoomPreferences()
+        resetMapToDefaults()
+    }
+
 }
