+++ app/src/main/java/pl/llp/aircasting/screens/new_session/confirmation/FixedSessionConfirmationViewMvcImpl.kt
@@ -3,7 +3,12 @@ package pl.llp.aircasting.screens.new_session.confirmation
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.fragment.app.FragmentManager
+import com.google.android.gms.maps.CameraUpdateFactory
+import com.google.android.gms.maps.GoogleMap
+import com.google.android.gms.maps.model.LatLng
+import com.google.android.gms.maps.model.MarkerOptions
 import pl.llp.aircasting.R
+import pl.llp.aircasting.lib.BitmapHelper
 import pl.llp.aircasting.models.Session
 
 class FixedSessionConfirmationViewMvcImpl(
@@ -24,4 +29,17 @@ class FixedSessionConfirmationViewMvcImpl(
     override fun updateLocation(latitude: Double?, longitude: Double?) {
         // do nothing, static map
     }
+
+    override fun onMapReady(googleMap: GoogleMap) {
+        mMap = googleMap
+        val sessionLocation = session?.location ?: return
+        val location = LatLng(sessionLocation.latitude, sessionLocation.longitude)
+        val icon = BitmapHelper.bitmapFromVector(context, R.drawable.ic_dot_20)
+        val marker = MarkerOptions()
+            .position(location)
+            .icon(icon)
+        mMarker = googleMap.addMarker(marker)
+        googleMap.moveCamera(CameraUpdateFactory.newLatLngZoom(location, DEFAULT_ZOOM))
+    }
+
 }
